Add openpyxl requirement and ignore *.xlsx files
The openpyxl library will be used to read xlsx files. Im using
this library instead of my normal xlrd (which can read both) for
this plugin as it is intended to only be used for xlsx files
anyways and I read that this library has better handling for large
xlsx files.  (I think it might be something like xlrd has to load
the entire file into memeory for xlsx files, but I'm not sure and
don't really care as long as this works well for large files).  If
support for both xls and xlsx files is needed, the planned two
seperate plugins will require both xlrd and openpyxl, but I think
this is fine under the premise of using the right tool for the
right job.

Added ignoring *.xlsx files to the .gitignore as we will probably
be using *.xlsx files for testing purposes.